+++ backend/expenses/Touir.ExpensesManager.Expenses/Controllers/Responses/ErrorResponse.cs
@@ -1,0 +1,8 @@
+namespace Touir.ExpensesManager.Expenses.Controllers.Responses
+{
+    public class ErrorResponse
+    {
+        public string? Message { get; set; }
+        public ErrorResponse() { }
+    }
+}